+++ algorithm/string/modules/test_hexdump.c
@@ -254,4 +254,3 @@ module_exit(test_hexdump_exit);
 
 MODULE_AUTHOR("Andy Shevchenko <andriy.shevchenko@linux.intel.com>");
 MODULE_LICENSE("Dual BSD/GPL");
-