+++ hpc/nvidia/cuda/fakeCUDA/utils.cpp
@@ -9,73 +9,8 @@
 #include <elf.h>
 #include "utils.hpp"
 #include "debug.h"
+#include "memshow.c"
 
-
-/**
- * like $ hexdump --canonical $FILE
- */
-void hexdump(const void *mem, size_t size)
-{
-	const int width = 16;
-	int nr_newline = 0;
-	int nr_startline = 0;
-	size_t align_size = 0;
-
-	assert(!(width % 8) && "width must align of 8");
-
-	while (align_size < size)
-		align_size += width;
-
-	for (size_t i = 0; i < size; i++) {
-		bool startline = i % width == 0;
-		bool newline = (i + 1) % width == 0;
-
-		if (startline) {
-			printf("%#016lx | ", (uint64_t)mem + width * nr_newline);
-		}
-
-		uint8_t u8 = *(uint8_t *)((uint8_t *)mem + i);
-		printf("%02x%s", u8, newline ? " |" : " ");
-
-		if ((i + 1) % 8 == 0 && !newline)
-			printf(" ");
-
-		/**
-		 * Display memory as character
-		 */
-		if (newline) {
-			const void *memch = (uint8_t *)mem + width * nr_newline;
-			for (size_t j = 0; j < width; j++) {
-				uint8_t c8 = *(uint8_t *)((uint8_t *)memch + j);
-				printf("%c", isprint(c8) ? c8 : '.');
-			}
-			printf("|\n");
-		}
-
-		if (newline)
-			nr_newline++;
-		if (startline)
-			nr_startline++;
-	}
-
-	if (align_size > size) {
-		for (size_t i = 0; i < align_size - size; i++)
-			printf("   ");
-		printf("|");
-
-		const void *memlastrow = (uint8_t *)mem + width * nr_newline;
-		for (size_t i = 0; i < width - (align_size - size); i++) {
-			uint8_t c8 = *(uint8_t *)((uint8_t *)memlastrow + i);
-			printf("%c", isprint(c8) ? c8 : '.');
-		}
-		for (size_t i = 0; i < align_size - size; i++)
-			printf(" ");
-		printf("|");
-	}
-
-	if (nr_newline != nr_startline)
-		printf("\n");
-}
 
 void debug_hexdump(const void *mem, size_t size)
 {