+++ hpc/nvidia/cuda/fakeCUDA/utils.hpp
@@ -3,12 +3,12 @@
 #pragma once
 #include <elf.h>
 #include <sys/types.h>
+#include "memshow.h"
 
 
 #define ARRAY_SIZE(arr)	(sizeof(arr) / sizeof(arr[0]))
 
 
-void hexdump(const void *mem, size_t size);
 void debug_hexdump(const void *mem, size_t size);
 
 void output_memory_to_file(const char *file, const void *mem, size_t size);