+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - glibc: test `__GLIBC_USE()`;
 - shell: add some commands from ostools;
 #### Changed
+- Move hexdump() from fakeCUDA to libs/memshow.c;
 - Move vfio into virt/;
 - make: Support NOSUBDIR definition;
 #### Deprecated